+++ packages/odyssey-react-mui/src/CustomizableAutocomplete.tsx
@@ -13,79 +13,11 @@
 import {
   Autocomplete as MuiAutocomplete,
   AutocompleteProps as MuiAutocompleteProps,
-  InputBase,
-  UseAutocompleteProps,
-  AutocompleteValue,
-  AutocompleteRenderInputParams,
 } from "@mui/material";
-import {
-  createContext,
-  FC,
-  forwardRef,
-  HTMLAttributes,
-  memo,
-  ReactElement,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-} from "react";
-import styled from "@emotion/styled";
-import { VariableSizeList, ListChildComponentProps } from "react-window";
-import _AutoSizer, {
-  Props as AutoSizerProps,
-  Size as AutoSizerSize,
-} from "react-virtualized-auto-sizer";
+import { memo } from "react";
 
 import { AutocompleteProps } from "./Autocomplete";
-
-// This is required to get around a react-types issue for "AutoSizer is not a valid JSX element."
-// @see https://github.com/bvaughn/react-virtualized/issues/1739#issuecomment-1291444246
-const AutoSizer = _AutoSizer as unknown as FC<AutoSizerProps>;
-
-import { Field } from "./Field";
-import {
-  ComponentControlledState,
-  useInputValues,
-  getControlState,
-} from "./inputUtils";
-
-type SetItemSize = (index: number, size: number) => void;
-
-const Row = ({
-  data,
-  index,
-  setItemSize,
-  style,
-}: ListChildComponentProps & { setItemSize: SetItemSize }) => {
-  const rowRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (rowRef.current) {
-      /**
-       * Checking for child height to workaround a bug where the clientHeight of the row isn't updated correctly
-       * @see here if you need to know more: https://github.com/bvaughn/react-window/issues/582#issuecomment-1883074908
-       */
-      const firstChild = rowRef.current.firstElementChild;
-      if (firstChild) {
-        const height = firstChild.clientHeight;
-        setItemSize(index, height);
-      }
-    }
-  }, [index, rowRef, setItemSize]);
-
-  /**
-   * react-window calculates the absolute positions of the list items, via an inline style, so
-   * we need to add it to each list item that is being rendered in the viewable list window.
-   * @see here if you need to know more: https://github.com/bvaughn/react-window?tab=readme-ov-file#why-is-my-list-blank-when-i-scroll
-   */
-  return (
-    <div key={`${index}-${style}}`} style={style} ref={rowRef}>
-      {data[index]}
-    </div>
-  );
-};
+import { useAutocomplete } from "./useAutocomplete";
 
 export type CustomizableAutocompleteProps<
   OptionType,
@@ -113,11 +45,6 @@ export type CustomizableAutocompleteProps<
     IsCustomValueAllowed
   >["renderTags"];
 };
-
-const ListboxContainer = styled.div`
-  width: 100%;
-  height: 100%;
-`;
 
 const CustomizableAutocomplete = <
   OptionType,
@@ -159,215 +86,40 @@ const CustomizableAutocomplete = <
   HasMultipleChoices,
   IsCustomValueAllowed
 >) => {
-  const controlledStateRef = useRef(
-    getControlState({
-      controlledValue: value,
-      uncontrolledValue: defaultValue,
-    }),
-  );
-
-  const isVirtualized = useRef(Boolean(isVirtualizedProp));
-  const defaultValueProp = useMemo<
-    | AutocompleteValue<
-        OptionType,
-        HasMultipleChoices,
-        undefined,
-        IsCustomValueAllowed
-      >
-    | undefined
-  >(() => {
-    if (hasMultipleChoices) {
-      if (value === undefined) {
-        return defaultValue;
-      }
-      return [] as AutocompleteValue<
-        OptionType,
-        HasMultipleChoices,
-        undefined,
-        IsCustomValueAllowed
-      >;
-    }
-    return value === undefined ? defaultValue : undefined;
-  }, [defaultValue, hasMultipleChoices, value]);
-
-  const valueProps = useInputValues({
-    defaultValue: defaultValueProp,
-    value: value,
-    controlState: controlledStateRef.current,
+  const {
+    inputValueProp,
+    isVirtualized,
+    onChange,
+    onInputChange,
+    renderInput,
+    valueProps,
+    VirtualizedListboxComponent,
+  } = useAutocomplete<OptionType, HasMultipleChoices, IsCustomValueAllowed>({
+    ariaDescribedBy,
+    defaultValue,
+    errorMessage,
+    errorMessageList,
+    hasMultipleChoices,
+    hint,
+    HintLinkComponent,
+    inputValue,
+    isFullWidth,
+    isOptional,
+    isVirtualized: isVirtualizedProp,
+    label,
+    name: nameOverride,
+    onChange: onChangeProp,
+    onInputChange: onInputChangeProp,
+    testId,
+    value,
   });
-
-  const inputValueProp = useMemo(() => {
-    if (controlledStateRef.current === ComponentControlledState.CONTROLLED) {
-      return { inputValue };
-    }
-    return undefined;
-  }, [inputValue]);
-
-  const renderInput = useCallback(
-    ({
-      InputLabelProps,
-      InputProps,
-      ...params
-    }: AutocompleteRenderInputParams) => (
-      <Field
-        ariaDescribedBy={ariaDescribedBy}
-        errorMessage={errorMessage}
-        errorMessageList={errorMessageList}
-        fieldType="single"
-        hasVisibleLabel
-        //@ts-expect-error htmlFor does not exist ont he InputLabelProps for autocomplete
-        id={InputLabelProps.htmlFor}
-        isFullWidth={isFullWidth}
-        hint={hint}
-        HintLinkComponent={HintLinkComponent}
-        label={label}
-        isOptional={isOptional}
-        renderFieldComponent={({
-          ariaDescribedBy,
-          id,
-          errorMessageElementId,
-          labelElementId,
-        }) => (
-          <InputBase
-            {...params}
-            {...InputProps}
-            inputProps={{
-              ...params.inputProps,
-              "aria-errormessage": errorMessageElementId,
-              "aria-labelledby": labelElementId,
-              "data-se": testId,
-            }}
-            aria-describedby={ariaDescribedBy}
-            id={id}
-            name={nameOverride ?? id}
-            required={!isOptional}
-          />
-        )}
-      />
-    ),
-    [
-      ariaDescribedBy,
-      errorMessage,
-      errorMessageList,
-      hint,
-      HintLinkComponent,
-      isFullWidth,
-      isOptional,
-      label,
-      nameOverride,
-      testId,
-    ],
-  );
-
-  const OuterListboxContext = createContext({});
-
-  const OuterListboxElementType = forwardRef<HTMLDivElement>((props, ref) => {
-    const outerProps = useContext(OuterListboxContext);
-    return <div ref={ref} {...props} {...outerProps} />;
-  });
-
-  const ListboxComponent = forwardRef<
-    HTMLDivElement,
-    HTMLAttributes<HTMLElement>
-  >(function (props, ref) {
-    const { children, ...other } = props;
-    const itemData: ReactElement[] = (children as ReactElement[]).flatMap(
-      (item: ReactElement & { children?: ReactElement[] }) =>
-        [item].concat(item.children || []),
-    );
-    const sizeMap = useMemo(() => new Map<number, number>(), []);
-    const virtualizedListRef = useRef<VariableSizeList>(null);
-
-    const setItemSize = useCallback<SetItemSize>(
-      (index, size) => {
-        virtualizedListRef?.current?.resetAfterIndex(0, true);
-        sizeMap.set(index, size);
-      },
-      [virtualizedListRef, sizeMap],
-    );
-
-    // The number of items (rows or columns) to render outside of the visible area for performance and scrolling reasons
-    const overscanRowCount = 8;
-
-    const getItemSize = useCallback(
-      (index: number) => sizeMap.get(index) || 45,
-      [sizeMap],
-    );
-
-    const renderWindow = useCallback(
-      ({ height, width }: AutoSizerSize) => (
-        <VariableSizeList
-          innerElementType="ul"
-          itemData={itemData}
-          itemCount={itemData.length}
-          // eslint-disable-next-line @typescript-eslint/no-unused-vars
-          itemSize={getItemSize}
-          height={height}
-          width={width}
-          ref={virtualizedListRef}
-          outerElementType={OuterListboxElementType}
-          overscanCount={overscanRowCount}
-        >
-          {({ data, index, style }) => (
-            <Row
-              data={data}
-              index={index}
-              style={style}
-              setItemSize={setItemSize}
-            />
-          )}
-        </VariableSizeList>
-      ),
-      [itemData, getItemSize, setItemSize, virtualizedListRef],
-    );
-
-    return (
-      <ListboxContainer ref={ref}>
-        <OuterListboxContext.Provider value={other}>
-          <AutoSizer>{renderWindow}</AutoSizer>
-        </OuterListboxContext.Provider>
-      </ListboxContainer>
-    );
-  });
-
-  const onChange = useCallback<
-    NonNullable<
-      UseAutocompleteProps<
-        OptionType,
-        HasMultipleChoices,
-        undefined,
-        IsCustomValueAllowed
-      >["onChange"]
-    >
-  >(
-    (event, value, reason, details) => {
-      onChangeProp?.(event, value, reason, details);
-    },
-    [onChangeProp],
-  );
-
-  const onInputChange = useCallback<
-    NonNullable<
-      UseAutocompleteProps<
-        OptionType,
-        HasMultipleChoices,
-        undefined,
-        IsCustomValueAllowed
-      >["onInputChange"]
-    >
-  >(
-    (event, value, reason) => {
-      onInputChangeProp?.(event, value, reason);
-    },
-    [onInputChangeProp],
-  );
 
   return (
     <MuiAutocomplete
       {...valueProps}
       {...inputValueProp}
       // conditionally provide the ListboxComponent if this needs to be virtualized
-      {...(isVirtualized.current && { ListboxComponent })}
+      {...(isVirtualized.current && { VirtualizedListboxComponent })}
       // AutoComplete is wrapped in a div within MUI which does not get the disabled attr. So this aria-disabled gets set in the div
       aria-disabled={isDisabled}
       disableCloseOnSelect={hasMultipleChoices}
